+++ package.json
@@ -23,8 +23,12 @@
     ]
   },
   "dependencies": {
+    "@hapi/boom": "^10.0.1",
     "apollo-server": "^3.12.0",
+    "bcrypt": "^5.0.1",
+    "express": "^4.18.2",
     "graphql": "^16.7.1",
+    "http-server": "^14.1.1",
     "uuid": "^9.0.0"
   }
 }
